--- src/main/java/com/jkmconfecciones/Integrador_app/controller/AdminControlador.java
@@ -1,15 +1,0 @@
-package com.jkmconfecciones.Integrador_app.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin")
-public class AdminControlador {
-
-    @GetMapping("/admin")
-    public String paginaAdmin() {
-        return "admin/admin";
-    }
-}
