+++ src/main/java/com/jkmconfecciones/Integrador_app/controller/ControllerInicio.java
@@ -11,23 +11,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class ControllerInicio {
 
 
-    @GetMapping("/admin")
-    public String adminPage(Model model, @AuthenticationPrincipal UserDetails userDetails) {
-        model.addAttribute("usuario", userDetails != null ? userDetails.getUsername() : "Administrador");
-        return "admin";
-    }
-
-    @GetMapping("/usuario")
-    public String usuarioPage(Model model, @AuthenticationPrincipal UserDetails userDetails) {
-        model.addAttribute("usuario", userDetails != null ? userDetails.getUsername() : "Usuario");
-        return "usuario";
-    }
-
-    @GetMapping("/pedidos")
-    public String pedidosPage(Model model) {
-        model.addAttribute("mensaje", "Accede desde el panel de pedidos");
-        return "pedidos";
-    }
 
     @GetMapping("/recuperar")
     public String recuperarPage() {
