+++ src/main/resources/templates/admin/admin.html
@@ -89,7 +89,7 @@
                 <i class="fas fa-tachometer-alt w-5 mr-3"></i>
                 <span>Panel</span>
             </a>
-            <a href="/pedidos" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:text-white hover:border-l-4 hover:border-jk-red transition-all">
+            <a href="/admin/pedidos" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:text-white hover:border-l-4 hover:border-jk-red transition-all">
                 <i class="fas fa-shopping-cart w-5 mr-3"></i>
                 <span>Pedidos</span>
             </a>
@@ -125,7 +125,7 @@
                 <i class="fas fa-clipboard-list w-5 mr-3"></i>
                 <span>Control de Inventario</span>
             </a>
-            <a href="#" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
+            <a href="/admin/cargaMasivaDatos" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
                 <i class="fas fa-download w-5 mr-3"></i>
                 <span>Carga Masiva de Datos</span>
             </a>
@@ -133,11 +133,11 @@
                 <i class="fas fa-shield-alt w-5 mr-3"></i>
                 <span>Registro de Auditoría de Seguridad</span>
             </a>
-            <a href="#" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
+            <a href="/admin/edicionProducto" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
                 <i class="fas fa-edit w-5 mr-3"></i>
                 <span>Edición de Productos</span>
             </a>
-            <a href="#" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
+            <a href="/admin/notificaciones" class="flex items-center px-5 py-3 text-blue-200 hover:bg-jk-sidebar-hover hover:border-l-4 hover:border-jk-red transition-all">
                 <i class="fas fa-bell w-5 mr-3"></i>
                 <span>Notificaciones</span>
             </a>
