+++ src/main/resources/templates/admin/pedidos.html
