+++ src/main/resources/templates/registroAuditoriaSeguridad.html
@@ -130,7 +130,7 @@
 <span class="material-icons mr-3">inventory</span>
                     Control de Inventario
                 </a>
-<a class="flex items-center px-4 py-2 text-gray-300 hover:bg-gray-700 hover:text-white rounded transition-colors duration-200" href="cargaMasivaDatos.html">
+<a class="flex items-center px-4 py-2 text-gray-300 hover:bg-gray-700 hover:text-white rounded transition-colors duration-200" href="admin/cargaMasivaDatos.html">
 <span class="material-icons mr-3">upload_file</span>
                     Carga Masiva de Datos
                 </a>
